board:2093 bugprone-implicit-widening 28 -> 0 · integer-division 13 -> 0 · incorrect-roundings 9 -> 0
THE FIXITS FOR THE FIRST ONE ARE WRONG AND WERE THROWN AWAY. clang-tidy offered
`static_cast<size_type>(2 * sides[side])` -- casting the RESULT, which is exactly
the implicit widening it is complaining about, and `size_type` is not even a name
in scope at namespace level in most of these files. All twenty-eight are done by
hand and widen the FIRST OPERAND, which is where the arithmetic has to happen:
`positionM[static_cast<size_t>(b) * 3]`, `size_t{512} * 1024`. That is the round's
second wrong fixit and board:2093 already carries the rule.

Every integer division was an exact count -- triangles from indices, points from
an interleaved coordinate run, a ring index from a flat cell number -- and the
division is right; what was missing is the NAME. `static_cast<double>(roofRun.size() / 3)`
is `const size_t roofTriangles = roofRun.size() / 3;` and then the cast, and the
ledger line reads as what it publishes.

The roundings are std::lround now. `(int)(x + 0.5)` is the classic rounding bug
-- it is wrong for negatives and it can jump a whole integer when x + 0.5 rounds
up in the addition -- and every one of the nine sites here happens to feed it a
non-negative value, which is luck rather than design. DrawKey::Of loses its
constexpr for it: std::lround is not constexpr before C++26 (measured), and where
the standard supplies the function it is the one used -- a hand-written nearest
beside it would be the second mechanism CLAUDE.md's SDL3 rule refuses. Nothing
evaluates DrawKey::Of in a constant expression.

OldTown renders a23e97f2; Venice renders 9c9a4758, which is board:2092's second
Venice and one the owner counts as correct for now. outshine/places: 9 PASS, 0 FAIL.

+++ src/base/spatial/Wayfinding.cpp
@@ -164,7 +164,7 @@ size_t Network::Cross() {
     for (int side = 0; side < 2; ++side) {
       const Way &way = Ways_[ways[side]];
       const uint32_t local = sides[side] - static_cast<uint32_t>(way.First);
-      const double fromLat = Points_[2 * sides[side]];
+      const double fromLat = Points_[2 * static_cast<size_t>(sides[side])];
       const double fromLon = Points_[2 * sides[side] + 1];
       const double toLat = Points_[2 * sides[side] + 2];
       const double toLon = Points_[2 * sides[side] + 3];
