+++ src/render/SceneRenderer.cpp
@@ -1336,8 +1336,8 @@ ReadState SceneRenderer::ReadKeptIndices(uint32_t &kept, uint32_t &batches) {
   if (read.FromBuffer(Device_.Get(), args, bytes) != ReadState::Ready) { return ReadState::Failed; }
   const auto *const held = static_cast<const uint32_t *>(static_cast<const void *>(read.Rows()));
   for (uint32_t at = 0; at < rows; ++at) {
-    kept += held[at * 5u];
-    batches += held[at * 5u] > 0u ? 1u : 0u;
+    kept += held[static_cast<size_t>(at) * 5];
+    batches += held[static_cast<size_t>(at) * 5] > 0u ? 1u : 0u;
   }
   return ReadState::Ready;
 }
